LPS-109946 Move to package .helper
